+++ packages/core/src/js/index.ts
@@ -85,5 +85,6 @@ export type { TimeToDisplayProps } from './tracing';
 
 export { Mask, Unmask } from './replay/CustomMask';
 
+export { FeedbackButton } from './feedback/FeedbackButton';
 export { FeedbackWidget } from './feedback/FeedbackWidget';
 export { showFeedbackWidget } from './feedback/FeedbackWidgetManager';