+++ samples/react-native/src/Screens/ErrorsScreen.tsx
@@ -238,6 +238,7 @@ const ErrorsScreen = (_props: Props) => {
         ) : null}
         <View style={styles.mainViewBottomWhiteSpace} />
       </ScrollView>
+      <Sentry.FeedbackButton/>
     </>
   );
 };
@@ -273,6 +274,7 @@ const styles = StyleSheet.create({
   },
   mainViewBottomWhiteSpace: {
     marginTop: 32,
+    marginBottom: 75,
   },
 });